Added the MaxMind GEOIP library
The MaxMind GEOIP library is a third party LGPL set of source files
which have been packaged into a separate library for use in portions of
the system which will identify the location of where an IP block is
assigned to.

No modifications have been made to this source code, the project files
where developed to build the source into a useable library.